fix dir & update .gitignore & del go.sum
