--- .vscode/launch.json
@@ -1,28 +1,0 @@
-{
-	// Use IntelliSense to learn about possible attributes.
-	// Hover to view descriptions of existing attributes.
-	// For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-	"version": "0.2.0",
-	"configurations": [
-		{
-			"name": "Launch Demo",
-			"type": "go",
-			"request": "launch",
-			"mode": "debug",
-			"program": "${workspaceFolder}/_demo",
-			"cwd": "${workspaceFolder}/_demo",
-			"env": {},
-			"args": []
-		},
-		{
-			"name": "Launch Scaffold",
-			"type": "go",
-			"request": "launch",
-			"mode": "debug",
-			"program": "${workspaceFolder}/_scaffold",
-			"cwd": "${workspaceFolder}/_scaffold",
-			"env": {},
-			"args": []
-		},
-	]
-}
