+++ _demo/app/fonts.go
@@ -1,4 +1,4 @@
-package main
+package app
 
 import (
 	"io/ioutil"
