+++ _demo/go.mod
@@ -9,5 +9,5 @@ require (
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0
 	github.com/hajimehoshi/ebiten/v2 v2.0.2
 	golang.org/x/image v0.0.0-20201208152932-35266b937fa6
-	golang.org/x/text v0.3.4
+	golang.org/x/text v0.3.0
 )
